Move DatabaseSettings to database module
