+++ pydantic_settings/settings.py
@@ -1,4 +1,3 @@
-import urllib.parse
 from inspect import getsourcefile
 from pathlib import Path
 from typing import Any, Callable, Dict, List, Optional, Pattern, Sequence, Tuple, Union
@@ -15,7 +14,7 @@ from pydantic.main import BaseModel
 from pydantic.networks import EmailStr, IPvAnyAddress
 from pydantic.types import FilePath
 
-from .database import DatabaseDsn
+from pydantic_settings.database import DatabaseSettings
 
 DEFAULT_SETTINGS_MODULE_FIELD = Field(
     "pydantic_settings.settings.PydanticSettings", env="DJANGO_SETTINGS_MODULE"
@@ -53,38 +52,6 @@ class SetUp(BaseSettings):
                 )
 
             settings.configure(**settings_dict)
-
-
-class DatabaseSettings(BaseSettings):
-    default: Optional[DatabaseDsn] = Field(env="DATABASE_URL")
-
-    @validator("*")
-    def format_database_settings(cls, v):
-        if v is None:
-            return {}
-
-        engines = {
-            "postgres": "django.db.backends.postgresql",
-            "postgis": "django.contrib.gis.db.backends.postgis",
-            "mssql": "sql_server.pyodbc",
-            "mysql": "django.db.backends.mysql",
-            "mysqlgis": "django.contrib.gis.db.backends.mysql",
-            "sqlite": "django.db.backends.sqlite3",
-            "spatialite": "django.contrib.gis.db.backends.spatialite",
-            "oracle": "django.db.backends.oracle",
-            "oraclegis": "django.contrib.gis.db.backends.oracle",
-            "redshift": "django_redshift_backend",
-        }
-
-        return {
-            "NAME": v.path[1:] if v.path.startswith("/") else v.path or "",
-            "USER": v.user or "",
-            "PASSWORD": v.password or "",
-            "HOST": urllib.parse.unquote(v.host) if v.host else "",
-            "PORT": v.port or "",
-            "CONN_MAX_AGE": 0,
-            "ENGINE": engines[v.scheme],
-        }
 
 
 class TemplateBackendModel(BaseModel):
